pyproject.toml
[project]
name = "minimal-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "mlflow>=2.19.0",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "polars>=1.19.0",
    "pyarrow>=18.1.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.0",
    "supabase>=2.11.0",
    "tomli>=2.2.1",
    "tqdm>=4.67.1",
]

[tool.hatch.build.targets.wheel]
packages = ["minimal_mlops"]
